modified:   Dockerfile
Changed the Docker version the Dockerfile was tested against.
Changed the naming convention of the Docker image to Paasta/Slalom.
Changed the parent Image to Paasta/Base.
Changed the Home Directory of the Image user from '/home/git' to '/root'
Removed anything pertaining to the 'git' user.
Removed the installation of Forever because it's now prepackaged with the base image.